+++ branches/rel-1-2-p2/opendocman/install/odm.php
@@ -26,7 +26,7 @@ DROP TABLE IF EXISTS admin
 
 $result = mysql_query("
 CREATE TABLE admin (
-  id smallint(5) unsigned default NULL,
+  id int(11) unsigned default NULL,
   admin tinyint(4) default NULL
 ) TYPE=MyISAM
 ") or die("<br>Could not create admin table" .  mysql_error());
@@ -43,7 +43,7 @@ DROP TABLE IF EXISTS category
 
 $result = mysql_query("
 CREATE TABLE category (
-  id smallint(5) unsigned NOT NULL auto_increment,
+  id int(11) unsigned NOT NULL auto_increment,
   name varchar(255) NOT NULL default '',
   PRIMARY KEY  (id)
 ) TYPE=MyISAM
@@ -72,9 +72,9 @@ DROP TABLE IF EXISTS data
 
 $result = mysql_query("
 CREATE TABLE data (
-  id smallint(5) unsigned NOT NULL auto_increment,
-  category tinyint(4) unsigned NOT NULL default '0',
-  owner smallint(6) unsigned default NULL,
+  id int(11) unsigned NOT NULL auto_increment,
+  category int(11) unsigned NOT NULL default '0',
+  owner int(11) unsigned default NULL,
   realname varchar(255) NOT NULL default '',
   created datetime NOT NULL default '0000-00-00 00:00:00',
   description varchar(255) default NULL,
@@ -83,7 +83,7 @@ CREATE TABLE data (
   department smallint(6) unsigned default NULL,
   default_rights tinyint(4) default NULL,
   publishable tinyint(4) default NULL,
-  reviewer smallint(6) unsigned default NULL,
+  reviewer int(11) unsigned default NULL,
   reviewer_comments varchar(255) default NULL,
   PRIMARY KEY  (id),
   KEY data_idx (id,owner),
@@ -101,7 +101,7 @@ DROP TABLE IF EXISTS department
 
 $result = mysql_query("
 CREATE TABLE department (
-  id smallint(5) unsigned NOT NULL auto_increment,
+  id int(11) unsigned NOT NULL auto_increment,
   name varchar(255) NOT NULL default '',
   PRIMARY KEY  (id)
 ) TYPE=MyISAM
@@ -118,8 +118,8 @@ DROP TABLE IF EXISTS dept_perms
 
 $result = mysql_query("
 CREATE TABLE dept_perms (
-  fid smallint(5) unsigned default NULL,
-  dept_id smallint(5) unsigned default NULL,
+  fid int(11) unsigned default NULL,
+  dept_id int(11) unsigned default NULL,
   rights tinyint(4) NOT NULL default '0',
   KEY rights (rights),
   KEY dept_id (dept_id),
@@ -134,8 +134,8 @@ DROP TABLE IF EXISTS dept_reviewer
 
 $result = mysql_query("
 CREATE TABLE dept_reviewer (
-  dept_id smallint(5) unsigned default NULL,
-  user_id smallint(5) unsigned default NULL
+  dept_id int(11) unsigned default NULL,
+  user_id int(11) unsigned default NULL
 ) TYPE=MyISAM
 ") or die("<br>Could not create dept_reviewer table");
 
@@ -151,7 +151,7 @@ DROP TABLE IF EXISTS log
 
 $result = mysql_query("
 CREATE TABLE log (
-  id int(10) unsigned NOT NULL default '0',
+  id int(11) unsigned NOT NULL default '0',
   modified_on datetime NOT NULL default '0000-00-00 00:00:00',
   modified_by varchar(25) default NULL,
   note text,
@@ -205,10 +205,10 @@ DROP TABLE IF EXISTS user
 
 $result = mysql_query("
 CREATE TABLE user (
-  id smallint(5) unsigned NOT NULL auto_increment,
+  id int(11) unsigned NOT NULL auto_increment,
   username varchar(25) NOT NULL default '',
   password varchar(50) NOT NULL default '',
-  department smallint(5) unsigned default NULL,
+  department int(11) unsigned default NULL,
   phone varchar(20) default NULL,
   Email varchar(50) default NULL,
   last_name varchar(255) default NULL,
@@ -229,8 +229,8 @@ DROP TABLE IF EXISTS user_perms
 
 $result = mysql_query("
 CREATE TABLE IF NOT EXISTS user_perms (
-  fid smallint(5) unsigned default NULL,
-  uid smallint(5) unsigned NOT NULL default '0',
+  fid int(11) unsigned default NULL,
+  uid int(11) unsigned NOT NULL default '0',
   rights tinyint(4) NOT NULL default '0',
   KEY user_perms_idx (fid,uid,rights),
   KEY fid (fid),
